Move detail view to bottom sheet
Instead of launching a new activity, the detail view is now in a
bottom sheet, which is a type of Material dialog that can be dragged
up and down the screen—more reminiscent of OmniFocus on iOS and more
Material-y. Plus it looks nicer and is a little easier to understand
and use, which is important for understanding when to save the user's
changes later.