Add date & time picker
We need to have some reliable way of picking dates/times for due
dates & defer dates. The best way to do that is a date picker
followed by a time picker, with a 'clear' button somewhere in the mix.
It mostly works (a little buggy), but handles everything it needs to.